job-init: avoid kill nothing
cat: can't open '/tmp/pid-tail': No such file or directory
sh: you need to specify whom to kill

+++ lib/job-init.sh
@@ -149,6 +149,7 @@ wait_on_manual_check()
 
 clean_job_resource()
 {
+	test -f /tmp/pid-tail || return
 	kill $(cat /tmp/pid-tail)
 }
